Petal shrinking on a timer / empty water / lose life
- Petals strink on a timer, although at a static rate
- There is empty space on the board; if the player falls into it, the game
- pauses in a dead state for a while.

There was a significant refactor of the display logic as the changes I needed to
make this work required a bunch of petal-specific logic.

There are a bunch of things I'm missing, because Elm requires a rearchitect to
handle randomness in a way that I don't currently have the time to handle properly.